+++ src/main/java/novamachina/exnihilosequentia/common/utility/ExNihiloConstants.java
@@ -110,6 +110,8 @@ public class ExNihiloConstants {
         public static final String SILKWORM = "silkworm";
         public static final String WITCH_WATER_BUCKET = "bucket_witchwater";
         public static final String YELLOW_CORAL_SEED = "seed_yellow_coral";
+        public static final String SEED_FERN = "fern";
+        public static final String SEED_LARGE_FERN = "large_fern";
 
         private Items() {
         }