+++ changelog.md
@@ -1,3 +1,6 @@
-2.0.1.0
-- Fix recipe conflicts
-- Fix incorrect textures for flint and iron meshes
+2.0.2.0
+- Fix crash on start issue
+- Add prismarine crystal drops to sieve
+- Add seeds to vanilla composter
+- Add fern seeds and beetroot drops to sieve
+- Add campfire and smoker silkworm recipes